Moved chat files chat package
